Relocated Footer.js to main
