+++ src/components/main/HomeScreen.js
@@ -19,7 +19,7 @@ import { FontAwesome, Ionicons } from '@expo/vector-icons'
 
 import Colors from '../../assets/Colors'
 
-import Footer from '../utilities/Footer.js'
+import Footer from './Footer.js'
 
 import Auth from '@aws-amplify/auth'
