+++ src/components/main/noFlight/ReceiptWithoutFlightScreen.js
@@ -15,7 +15,7 @@ import Dash from 'react-native-dash';
 
 import Colors from '../../../assets/Colors.js';
 
-import Footer from '../../utilities/Footer.js';
+import Footer from '../Footer.js';
 
 import Auth from '@aws-amplify/auth';
