+++ src/components/main/payment/ThankYouScreen.js
@@ -15,7 +15,7 @@ import { FontAwesome, Feather } from '@expo/vector-icons';
 
 import Colors from '../../../assets/Colors.js';
 
-import Footer from '../../utilities/Footer.js';
+import Footer from '../Footer.js';
 
 import Auth from '@aws-amplify/auth';
